Created 'real' notification template
While filling the template, various things had to be done.

The email-context had to be filled with information. Most notable this is
some way to provide valid links for the emails. Because the app should
not rely on 'django.contrib.sites' and there is no 'Request'-object
available while sending mails, the necessary values have to be provided
with app-specific settings.

This might be something, that is not really user-friendly, but on the
other hand: admins might just hardcode some variables in their mail
templates.