Prepared admin integration
This is basically the start of the development of admin-related functions
of django-auth_enhanced.

Development will be split between modifications of the list-view inside
Django's admin backend and the actual detail view of a single User
object.

However, this will be the branch that represents the pull request that
is going to close #4.